// UaRestGateway/uarestgateway.client/src/opcua/models/Variant.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 1.05.4
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Variant
 */
export interface Variant {
    /**
     * 
     * @type {number}
     * @memberof Variant
     */
    Type?: number;
    /**
     * 
     * @type {any}
     * @memberof Variant
     */
    Body?: any | null;
    /**
     * 
     * @type {Array<number>}
     * @memberof Variant
     */
    Dimensions?: Array<number>;
}

/**
 * Check if a given object implements the Variant interface.
 */
export function instanceOfVariant(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function VariantFromJSON(json: any): Variant {
    return VariantFromJSONTyped(json, false);
}

export function VariantFromJSONTyped(json: any, ignoreDiscriminator: boolean): Variant {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'Type': !exists(json, 'Type') ? undefined : json['Type'],
        'Body': !exists(json, 'Body') ? undefined : json['Body'],
        'Dimensions': !exists(json, 'Dimensions') ? undefined : json['Dimensions'],
    };
}

export function VariantToJSON(value?: Variant | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Type': value.Type,
        'Body': value.Body,
        'Dimensions': value.Dimensions,
    };
}